OpenDCL 6.0 RC1
git-svn-id: https://svn.code.sf.net/p/opendcl/code/Trunk@273 07a8c159-f786-4120-a571-e88efff3e5ae

+++ Common/Diagnostic.h
@@ -325,7 +325,7 @@ template<>
 inline
 const TCHAR* asString( const std::vector< CString >& rStr )
 {
-	unsigned int ctElem = rStr.size();
+	size_t ctElem = rStr.size();
 	if( ctElem == 0 )
 		return _T("<empty>");
 	static TCHAR buf[1024];
